refactor(napi): move native transformer into Rust
