+++ crates/ox_content_napi/src/lib.rs
@@ -6,6 +6,7 @@
 mod mdast;
 mod mdast_raw;
 mod transfer;
+mod transformer;
 
 use napi::bindgen_prelude::*;
 use napi::Task;
@@ -13,11 +14,11 @@ use napi_derive::napi;
 use std::collections::HashMap;
 
 use ox_content_allocator::Allocator;
-use ox_content_ast::{Document, Heading, Node};
 use ox_content_parser::{Parser, ParserOptions};
-use ox_content_renderer::{HtmlRenderer, HtmlRendererOptions};
+use ox_content_renderer::HtmlRenderer;
 use ox_content_search::{DocumentIndexer, SearchIndex, SearchIndexBuilder, SearchOptions};
 use transfer::TransferPayloadKind;
+use transformer::{parse_frontmatter, MarkdownTransformer};
 
 /// Parse result containing the AST as JSON.
 #[napi(object)]
@@ -230,202 +231,7 @@ pub fn version() -> String {
 #[napi]
 pub fn transform(source: String, options: Option<JsTransformOptions>) -> TransformResult {
     let opts = options.unwrap_or_default();
-    let toc_max_depth = opts.toc_max_depth.unwrap_or(3);
-
-    // Parse frontmatter
-    let (content, frontmatter) = parse_frontmatter(&source);
-
-    // Parse markdown
-    let allocator = Allocator::new();
-    let parser_options = transform_options_to_parser_options(&opts);
-    let parser = Parser::with_options(&allocator, &content, parser_options);
-
-    let result = parser.parse();
-    match result {
-        Ok(doc) => {
-            // Extract TOC from headings
-            let toc = extract_toc(&doc, toc_max_depth);
-
-            // Render to HTML
-            let renderer_options = transform_options_to_renderer_options(&opts);
-            let mut renderer = HtmlRenderer::with_options(renderer_options);
-            let html = renderer.render(&doc);
-
-            TransformResult {
-                html,
-                frontmatter: serde_json::to_string(&frontmatter)
-                    .unwrap_or_else(|_| "{}".to_string()),
-                toc,
-                errors: vec![],
-            }
-        }
-        Err(e) => TransformResult {
-            html: String::new(),
-            frontmatter: "{}".to_string(),
-            toc: vec![],
-            errors: vec![e.to_string()],
-        },
-    }
-}
-
-/// Parses YAML frontmatter from Markdown content.
-fn parse_frontmatter(source: &str) -> (String, HashMap<String, serde_json::Value>) {
-    let mut frontmatter = HashMap::new();
-
-    // Check for frontmatter delimiter
-    if !source.starts_with("---") {
-        return (source.to_string(), frontmatter);
-    }
-
-    // Find the closing delimiter
-    let rest = &source[3..];
-    let Some(end_pos) = rest.find("\n---") else {
-        return (source.to_string(), frontmatter);
-    };
-
-    let frontmatter_str = &rest[..end_pos].trim_start_matches('\n');
-    let content = &rest[end_pos + 4..].trim_start_matches('\n');
-
-    // Parse simple YAML key-value pairs
-    for line in frontmatter_str.lines() {
-        let line = line.trim();
-        if line.is_empty() || line.starts_with('#') {
-            continue;
-        }
-
-        if let Some(colon_pos) = line.find(':') {
-            let key = line[..colon_pos].trim().to_string();
-            let value_str = line[colon_pos + 1..].trim();
-
-            let value = if value_str == "true" {
-                serde_json::Value::Bool(true)
-            } else if value_str == "false" {
-                serde_json::Value::Bool(false)
-            } else if let Ok(n) = value_str.parse::<i64>() {
-                serde_json::Value::Number(n.into())
-            } else if let Ok(n) = value_str.parse::<f64>() {
-                serde_json::Number::from_f64(n).map_or_else(
-                    || serde_json::Value::String(value_str.to_string()),
-                    serde_json::Value::Number,
-                )
-            } else {
-                // Remove surrounding quotes if present
-                let s = value_str.trim_matches('"').trim_matches('\'');
-                serde_json::Value::String(s.to_string())
-            };
-
-            frontmatter.insert(key, value);
-        }
-    }
-
-    (content.to_string(), frontmatter)
-}
-
-/// Extracts table of contents from document headings.
-fn extract_toc(doc: &Document, max_depth: u8) -> Vec<TocEntry> {
-    let mut entries = Vec::new();
-
-    for node in &doc.children {
-        if let Node::Heading(heading) = node {
-            if heading.depth <= max_depth {
-                let text = extract_heading_text(heading);
-                let slug = slugify(&text);
-                entries.push(TocEntry { depth: heading.depth, text, slug });
-            }
-        }
-    }
-
-    entries
-}
-
-/// Extracts plain text from a heading node.
-fn extract_heading_text(heading: &Heading) -> String {
-    let mut text = String::new();
-    for child in &heading.children {
-        collect_text(child, &mut text);
-    }
-    text
-}
-
-/// Recursively collects text from nodes.
-fn collect_text(node: &Node, text: &mut String) {
-    match node {
-        Node::Text(t) => text.push_str(t.value),
-        Node::Emphasis(e) => {
-            for child in &e.children {
-                collect_text(child, text);
-            }
-        }
-        Node::Strong(s) => {
-            for child in &s.children {
-                collect_text(child, text);
-            }
-        }
-        Node::InlineCode(c) => text.push_str(c.value),
-        Node::Delete(d) => {
-            for child in &d.children {
-                collect_text(child, text);
-            }
-        }
-        Node::Link(l) => {
-            for child in &l.children {
-                collect_text(child, text);
-            }
-        }
-        _ => {}
-    }
-}
-
-/// Converts text to URL-friendly slug.
-fn slugify(text: &str) -> String {
-    text.to_lowercase()
-        .chars()
-        .map(|c| if c.is_alphanumeric() || c == ' ' || c == '-' { c } else { ' ' })
-        .collect::<String>()
-        .split_whitespace()
-        .collect::<Vec<_>>()
-        .join("-")
-}
-
-/// Converts transform options to parser options.
-fn transform_options_to_parser_options(opts: &JsTransformOptions) -> ParserOptions {
-    let mut options =
-        if opts.gfm.unwrap_or(false) { ParserOptions::gfm() } else { ParserOptions::default() };
-
-    if let Some(v) = opts.footnotes {
-        options.footnotes = v;
-    }
-    if let Some(v) = opts.task_lists {
-        options.task_lists = v;
-    }
-    if let Some(v) = opts.tables {
-        options.tables = v;
-    }
-    if let Some(v) = opts.strikethrough {
-        options.strikethrough = v;
-    }
-    if let Some(v) = opts.autolinks {
-        options.autolinks = v;
-    }
-
-    options
-}
-
-/// Converts transform options to renderer options.
-fn transform_options_to_renderer_options(opts: &JsTransformOptions) -> HtmlRendererOptions {
-    let mut options = HtmlRendererOptions::new();
-
-    if let Some(v) = opts.convert_md_links {
-        options.convert_md_links = v;
-    }
-    if let Some(ref v) = opts.base_url {
-        options.base_url.clone_from(v);
-    }
-    if let Some(ref v) = opts.source_path {
-        options.source_path.clone_from(v);
-    }
-
-    options
+    MarkdownTransformer::from_options(&opts).transform(&source)
 }
 
 // =============================================================================
@@ -483,39 +289,7 @@ impl Task for TransformTask {
     type JsValue = TransformResult;
 
     fn compute(&mut self) -> Result<Self::Output> {
-        let toc_max_depth = self.options.toc_max_depth.unwrap_or(3);
-
-        // Parse frontmatter
-        let (content, frontmatter) = parse_frontmatter(&self.source);
-
-        // Parse markdown
-        let allocator = Allocator::new();
-        let parser_options = transform_options_to_parser_options(&self.options);
-        let parser = Parser::with_options(&allocator, &content, parser_options);
-
-        let result = match parser.parse() {
-            Ok(doc) => {
-                let toc = extract_toc(&doc, toc_max_depth);
-                let renderer_options = transform_options_to_renderer_options(&self.options);
-                let mut renderer = HtmlRenderer::with_options(renderer_options);
-                let html = renderer.render(&doc);
-
-                TransformResult {
-                    html,
-                    frontmatter: serde_json::to_string(&frontmatter)
-                        .unwrap_or_else(|_| "{}".to_string()),
-                    toc,
-                    errors: vec![],
-                }
-            }
-            Err(e) => TransformResult {
-                html: String::new(),
-                frontmatter: "{}".to_string(),
-                toc: vec![],
-                errors: vec![e.to_string()],
-            },
-        };
-        Ok(result)
+        Ok(MarkdownTransformer::from_options(&self.options).transform(&self.source))
     }
 
     fn resolve(&mut self, _env: Env, output: Self::Output) -> Result<Self::JsValue> {
